Python_Code/pyproject.toml
[project]
name = "theobserver"
version = "0.1.0"
description = ""
authors = [
    {name = "FHTW", email = "if20b219@technikum-wien.at"},
]
dependencies = [
    "paho-mqtt",
    "opencv-python",
    "imutils",
    "pyyaml",
    "requests",
    "flask>=2.2.2",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[tool]
[tool.pdm]
[[tool.pdm.source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[tool.pdm.dev-dependencies]
dev = []
